+++ core/src/main/java/me/mykindos/betterpvp/core/components/clans/data/ClanEnemy.java
@@ -7,18 +7,18 @@ import me.mykindos.betterpvp.core.components.clans.IClan;
 public class ClanEnemy {
 
     private final IClan clan;
-    private int dominance;
+    private double dominance;
 
-    public ClanEnemy(IClan clan, int dominance) {
+    public ClanEnemy(IClan clan, double dominance) {
         this.clan = clan;
         this.dominance = dominance;
     }
 
-    public void addDominance(int dominance) {
+    public void addDominance(double dominance) {
         this.dominance = Math.min(100, this.dominance + dominance);
     }
 
-    public void takeDominance(int dominance) {
+    public void takeDominance(double dominance) {
         this.dominance = Math.max(0, this.dominance - dominance);
     }
